optimized prop types structure
