+++ src/proptypes/card-prop-types.js
@@ -1,0 +1,38 @@
+import PropTypes from "prop-types";
+
+const cardPropTypes = {
+  bedrooms: PropTypes.number.isRequired,
+  city: PropTypes.shape({
+    location: PropTypes.shape({
+      latitude: PropTypes.number.isRequired,
+      longitude: PropTypes.number.isRequired,
+      zoom: PropTypes.number.isRequired,
+    }),
+    name: PropTypes.string.isRequired,
+  }),
+  description: PropTypes.string.isRequired,
+  goods: PropTypes.arrayOf(PropTypes.string.isRequired),
+  host: PropTypes.shape({
+    avatarUrl: PropTypes.string.isRequired,
+    id: PropTypes.number.isRequired,
+    isPro: PropTypes.bool.isRequired,
+    name: PropTypes.string.isRequired,
+  }),
+  id: PropTypes.number.isRequired,
+  images: PropTypes.arrayOf(PropTypes.string.isRequired),
+  isFavorite: PropTypes.bool.isRequired,
+  isPremium: PropTypes.bool.isRequired,
+  location: PropTypes.shape({
+    latitude: PropTypes.number.isRequired,
+    longitude: PropTypes.number.isRequired,
+    zoom: PropTypes.number.isRequired,
+  }),
+  maxAdults: PropTypes.number.isRequired,
+  previewImage: PropTypes.string.isRequired,
+  price: PropTypes.number.isRequired,
+  rating: PropTypes.number.isRequired,
+  title: PropTypes.string.isRequired,
+  type: PropTypes.string.isRequired,
+};
+
+export default cardPropTypes;
